requirements.txt
discord.py
SQLAlchemy
aiomysql
python-dotenv
cryptography